requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
Jinja2>=3.1.0
cloudscraper>=1.2.71
Pillow>=10.0.0
